Bug 1597904 - Create the main background script.
I was waiting for bug 1599360 and bug 1596258 to be resolved
first before asking for review. My wip branch is based on those
bugs. But we can start review on most of the background script
and some of the tests without them. Those bugs make the
additional `UPDATE_ASK` and `UPDATE_REFRESH` tips possible, plus
an update check from within the extension. I'll file follow-up
bugs for those additional tips once this is reviewed and merged,
and depending on the outcomes of the two bugs.

The tests are based on the update tests in
toolkit/mozapps/update/tests/browser and use the same helpers.