Bug 1604692 - Don't open the survey page once the user has started it.
We can use a content script in the survey page and listen for the
submit event. The code comments in survey.js explain a little
more. For reference, the survey URL is
https://qsurvey.mozilla.com/s3/Search-Interventions

This also bumps the add-on version since this is the first change
since I last asked the add-on to be signed (today).