Bug 1605351 - Don't remove all storage when showing the survey on startup; other storage and survey-on-startup changes.
The reason this bug happens is that I clear the entire storage
after opening the survey on startup, so the `surveyOpenedCount`
is forgotten. I should only remove the flag that indicates we
need to open the survey on startup.

This bug plus some prior conversation with QA made me realize
that the add-on doesn't handle the survey for the refresh and
update_refresh tips very well. It treats those tips as two of the
cases where we should open the survey on startup, since
refreshing the browser restarts it. But of course once the
browser restarts due to refresh, all add-ons are uninstalled,
including this one, so it can't possibly open the survey then. So
this patch opens the survey immediately for these two tips. The
user can at least see the survey in these cases if they cancel
the refresh dialog.

Other changes:

* Rename `storage.surveyPickedTip` to `surveyToOpenOnStartup` for
  clarity.
* Simplify some other uses of `browser.storage.local.set`. `set`
  doesn't actually set the whole storage to the object you pass
  in. Instead it merges the object into the storage. So there's
  no need to first get the storage, update a single property, and
  then set it. You can just set it.
* Bump the version to 1.0a5 since this is the first change since
  the add-on was last signed.